# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/documentation/files": {
      "get": {
        "operationId": "DocumentationController_getOutputFiles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Files retrieved successfully"
          },
          "404": {
            "description": "Output directory not found"
          }
        },
        "summary": "Get list of generated documentation files",
        "tags": [
          "Documentation"
        ]
      }
    },
    "/documentation/file": {
      "get": {
        "operationId": "DocumentationController_getFileContent",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get content of a documentation file",
        "tags": [
          "Documentation"
        ]
      },
      "delete": {
        "operationId": "DocumentationController_deleteFile",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a documentation file",
        "tags": [
          "Documentation"
        ]
      }
    },
    "/documentation/generate": {
      "post": {
        "operationId": "DocumentationController_generateDocumentation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Generate a Markdown documentation file for a project folder",
        "tags": [
          "Documentation"
        ]
      }
    },
    "/documentation/upload": {
      "post": {
        "operationId": "DocumentationController_uploadAndGenerateDocs",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Upload ZIP file and generate docs",
        "tags": [
          "Documentation"
        ]
      }
    },
    "/file-manager/list": {
      "get": {
        "operationId": "FileManagerController_listFiles",
        "parameters": [
          {
            "name": "folder",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FileManager"
        ]
      }
    },
    "/file-manager/content": {
      "get": {
        "operationId": "FileManagerController_getFileContent",
        "parameters": [
          {
            "name": "file",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FileManager"
        ]
      }
    },
    "/file-manager/file": {
      "delete": {
        "operationId": "FileManagerController_deleteFile",
        "parameters": [
          {
            "name": "file",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FileManager"
        ]
      }
    },
    "/file-manager/folder": {
      "delete": {
        "operationId": "FileManagerController_deleteFolder",
        "parameters": [
          {
            "name": "folder",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FileManager"
        ]
      }
    },
    "/file-manager/rename": {
      "patch": {
        "operationId": "FileManagerController_rename",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FileManager"
        ]
      }
    }
  },
  "info": {
    "title": "API Documentation",
    "description": "The API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateDocumentationDto": {
        "type": "object",
        "properties": {}
      },
      "RenameDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}